backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2
pillow==10.3.0
python-multipart==0.0.9
redis==5.0.4
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
coverage==7.5.1
